Refactoring simulation methods into optimization.py
