+++ coralme/core/__init__.py
@@ -3,3 +3,4 @@ import coralme.core.model
 import coralme.core.parameters
 import coralme.core.processdata
 import coralme.core.reaction
+import coralme.core.optimization
